pyproject.toml
[tool.poetry]
name = "cook-import"
version = "0.1.0"
description = ""
authors = ["Alexey Dubovskoy <dubovskoy.a@gmail.com>"]
packages = [{include = "cook_import", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10"
recipe-scrapers = ">=15.0.0"
pyinstaller = "*"
ingredient_parser_nlp = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[tool.poetry.scripts]
cook-import = 'cook_import.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"